+++ lib/commercetools-api/src/Models/Customer/CustomerDraftBuilder.php
@@ -192,6 +192,12 @@ final class CustomerDraftBuilder implements Builder
     private $authenticationMode;
 
     /**
+
+     * @var ?CustomerGroupAssignmentDraftCollection
+     */
+    private $customerGroupAssignments;
+
+    /**
      * <p>User-defined unique identifier for the Customer.
      * The <code>key</code> field is preferred over <code>customerNumber</code> as it is mutable and provides more flexibility.</p>
      *
@@ -507,6 +513,17 @@ final class CustomerDraftBuilder implements Builder
     }
 
     /**
+     * <p>Customer Groups to assign to the Customer.</p>
+     *
+
+     * @return null|CustomerGroupAssignmentDraftCollection
+     */
+    public function getCustomerGroupAssignments()
+    {
+        return $this->customerGroupAssignments;
+    }
+
+    /**
      * @param ?string $key
      * @return $this
      */
@@ -804,6 +821,17 @@ final class CustomerDraftBuilder implements Builder
     }
 
     /**
+     * @param ?CustomerGroupAssignmentDraftCollection $customerGroupAssignments
+     * @return $this
+     */
+    public function withCustomerGroupAssignments(?CustomerGroupAssignmentDraftCollection $customerGroupAssignments)
+    {
+        $this->customerGroupAssignments = $customerGroupAssignments;
+
+        return $this;
+    }
+
+    /**
      * @deprecated use withAnonymousCart() instead
      * @return $this
      */
@@ -865,7 +893,8 @@ final class CustomerDraftBuilder implements Builder
             $this->locale,
             $this->salutation,
             $this->stores,
-            $this->authenticationMode
+            $this->authenticationMode,
+            $this->customerGroupAssignments
         );
     }