+++ lib/commercetools-api/src/Models/Customer/CustomerGroupAssignment.php
@@ -1,0 +1,31 @@
+<?php
+
+declare(strict_types=1);
+/**
+ * This file has been auto generated
+ * Do not change it.
+ */
+
+namespace Commercetools\Api\Models\Customer;
+
+use Commercetools\Api\Models\CustomerGroup\CustomerGroupReference;
+use Commercetools\Base\DateTimeImmutableCollection;
+use Commercetools\Base\JsonObject;
+
+interface CustomerGroupAssignment extends JsonObject
+{
+    public const FIELD_CUSTOMER_GROUP = 'customerGroup';
+
+    /**
+     * <p>Reference to a Customer Group.</p>
+     *
+
+     * @return null|CustomerGroupReference
+     */
+    public function getCustomerGroup();
+
+    /**
+     * @param ?CustomerGroupReference $customerGroup
+     */
+    public function setCustomerGroup(?CustomerGroupReference $customerGroup): void;
+}