+++ lib/commercetools-api/src/Models/Customer/CustomerGroupAssignmentBuilder.php
@@ -1,0 +1,75 @@
+<?php
+
+declare(strict_types=1);
+/**
+ * This file has been auto generated
+ * Do not change it.
+ */
+
+namespace Commercetools\Api\Models\Customer;
+
+use Commercetools\Api\Models\CustomerGroup\CustomerGroupReference;
+use Commercetools\Api\Models\CustomerGroup\CustomerGroupReferenceBuilder;
+use Commercetools\Base\Builder;
+use Commercetools\Base\DateTimeImmutableCollection;
+use Commercetools\Base\JsonObject;
+use Commercetools\Base\JsonObjectModel;
+use Commercetools\Base\MapperFactory;
+use stdClass;
+
+/**
+ * @implements Builder<CustomerGroupAssignment>
+ */
+final class CustomerGroupAssignmentBuilder implements Builder
+{
+    /**
+
+     * @var null|CustomerGroupReference|CustomerGroupReferenceBuilder
+     */
+    private $customerGroup;
+
+    /**
+     * <p>Reference to a Customer Group.</p>
+     *
+
+     * @return null|CustomerGroupReference
+     */
+    public function getCustomerGroup()
+    {
+        return $this->customerGroup instanceof CustomerGroupReferenceBuilder ? $this->customerGroup->build() : $this->customerGroup;
+    }
+
+    /**
+     * @param ?CustomerGroupReference $customerGroup
+     * @return $this
+     */
+    public function withCustomerGroup(?CustomerGroupReference $customerGroup)
+    {
+        $this->customerGroup = $customerGroup;
+
+        return $this;
+    }
+
+    /**
+     * @deprecated use withCustomerGroup() instead
+     * @return $this
+     */
+    public function withCustomerGroupBuilder(?CustomerGroupReferenceBuilder $customerGroup)
+    {
+        $this->customerGroup = $customerGroup;
+
+        return $this;
+    }
+
+    public function build(): CustomerGroupAssignment
+    {
+        return new CustomerGroupAssignmentModel(
+            $this->customerGroup instanceof CustomerGroupReferenceBuilder ? $this->customerGroup->build() : $this->customerGroup
+        );
+    }
+
+    public static function of(): CustomerGroupAssignmentBuilder
+    {
+        return new self();
+    }
+}